win32: Fix symlink and junction support
Add support for CreateSymbolicLinkA and CreateSymbolicLinkW.
Seems there are again two types of calls one for wide strings and one
for ANSI ones. We also updated our CreateJunction function to be able
to handle wide strings instead of ANSI ones. We only implemented this
CreateJunction now for wide strings as junctions points are only
available for the public in Vista and later (and anything older is not
supported anyway anymore) why waste time on implementing an ANSI
version of it.

Fixes #575: charset problem in symlinks/junctions windows restore