ndmp: Fix calculation of ta->mover_state.record_num
Amanda has changed this calculation as they seem to have run into some
problem with tape writing. So lets import this change. We currently
probably don't run into this problem as we only emulate one big tape but
if we add support for real NDMP storage we might also be bitten by it.