ndmp: Fix DiskToDiskToTape support for NDMP.
We need to map the paired read storage of a NDMP backup to point to the
actual real storage before we start the migration or copy as otherwise
the Job tries to use the original storage of the NDMP Job and that is a
storage of the NDMP type which will trigger the assert in the code that
checks if the storage has the right password encoding.

We now check in the routines the actual JobType e.g. backup means write
storage anything else means read storage (e.g. restore, migration, copy)
We always try setting up paired read storage when we encounter NDMP
storage its not a backup Job. The has_paired_storage() function can now
also be used to determine if a Job has all NDMP paired storage so we
know in the copy/migrate code we need to setup paired storage. We can
call free_paired_storage() always as it checks if any paired storage was
setup using set_paired_storage().