socket: Introduce MaximumConnections for filed and stored.
When refactoring the socket_server code for the director we added a new
config option named MaximumConnections which is used to size the listen
backlog. This patch introduces something similar for the filed and
stored so we always size the backlog in a proper way according to the
wanted concurrency. We now check in accepting a connection in the filed
or stored if we already reached the wanted concurrency and then just
disconnect the session.

Fixes #593: Problem with a job on a passive client and limited Maximum
            Concurrent Jobs