python: Don't set python_loaded to true to early.
We only should set p_ctx->python_loaded to true when we found the module
and loaded it into the interpreter. As after setting p_ctx->python_loaded
we asume that we have a dictionary pointer to the python methods which
we don't as that only gets initialized after we actual resolve the
module name. We will crash on deferencing the dictionary pointer in any
subsequent call.