python: parse_plugin_definition consumes options.
From now on the parse_plugin_definition() function in all python plugins
will consume the options it needs for loading the plugin into the python
interpreter e.g. module_path and module_name and only pass all other
options to the python plugin definition parser.

For the python-fd this patch also adds the ability to push options to
the plugin without specifying module_path and module_name which is then
stored and prepended to the option string when we get an
bEventPluginCommand or bEventRestoreObject plugin event which
instantiates the plugin entry point.