Added Client Initiated Connection feature
This feature allows a Client (File Daemon) to open
and keep a connection to the Director.
It identifies himself and do the authentication.
The Director will keep the connection in a waiting pool.
If a job starts that needs access to the client
and a waiting connection to this client exist,
the connection will be used and removed from the pool.
The Client notices that this connection is now used
and opens a additional one connection
(required for multiple simultaneous client commands,
e.g. cancel a running job).

Configuration to enable Client Initiated Connections:
      bareos-dir.conf:
        Client
          ConnectionFromDirectorToClient = no
          ConnectionFromClientToDirector = yes

      bareos-fd.conf:
        Director
          Address
          ConnectionFromDirectorToClient = no
          ConnectionFromClientToDirector = yes

The "no" settings are not required, but gurantees that the
network connection will be only opened in the intended direction.

It is possible, to use Director Initiated Connections (default) and
Client Initiated Connection simultaneously. If a waiting connection
exists, it will be used, otherwise the Director tries to connect to the
client.

This patch also implements TCP Keepalive for Windows.

Fixes #455: Connect from File Daemon to Director Daemon