Restore to local VMDK File and Snapshot Naming.
When passing the option localvmdk=yes the plugin will run
bareos_vadp_dumper to restore to local VMDK file(s) instead
of restoring the original VM. In this case, the replace modes "Always"
and "Never" are handled properly.

The Snapshots that are created on backup now contain the JobId in the
snapshot name and both jobId and JobName in the snapshot description.

Also temporary files that are created on backup and restore are now
cleaned up.