Work around VDDK exit handling.
We always get an exit code of 0 although we exit with other values
when we are in error.

It seems the VDDK registers an atexit(3) handler. This code adds our own
atexit handler and as things are called in reverse order we should be
the first one called.

It also seems with older VDDKs it even more broken so we now call _exit
and hope that by bypassing the VDDK cleanup we don't break to much.