dird: Make storage more generic.
As we eventually want to support different types of storage e.g. Native
and NDMP to start with we need to make things somewhat more generic.
This is the first step by moving some of the generic storage functions
of the director into a new file storage.c which will try to abstract the
different storage types and have a generic abstraction so the rest of
the code needs as little change as possible.