ceph: Add support for rados_create2() API.
We currently use the rados_create() API call but there is also
the more flexible rados_create2() API. This patch adds support
for this other API. Based on the LIBRADOS_VERSION_CODE we
determine if the API is available. We check if we are being
compiled on a platform with a API version lower then 0.68.0
as that is the first version where the new usable version of
rados_create2 with the 4 flags is available. This was verified
by checking the git repo of CEPH and looking for the moment the
new API calls was extended to the 4 argument version and then
searching back in the history for the first bump of the
LIBRADOS_VER_MINOR or LIBRADOS_VER_EXTRA.