dird: Interactive run doesn't update pool on level change.
In interactive mode when changing the level we correctly call
apply_pool_overrides() which sets the new pool right but then we
call reset_restore_context() which sets the pool back to rc.pool
(e.g. restore context pool definition which still points to the original
 pool defined in the Job resource.)

This patch updates the rc.pool to point to the pool which was set by
the call to apply_pool_overrides() and set a variable in the restore
context which keeps track that we got a level override so we don't
overwrite the pool_source which is also set by apply_pool_overrides()
in reset_restore_context().

When we get a real pool override via the mod pool option we reset this
level override flag so we set the right pool_source in
reset_restore_context() again.

Fixes #633: Interactive run doesn't update pool on level change.