Fix problems with wide char symlinks and junctions
There were some problems since the last change to
the junctions and symlink code on windows to support
wide characters.

This fixes these problems so that we now can create correct
wide character symlinks and junctions.

This commit also adds a unit test for creating junctions and
changes what is needed  for that.

Fixes #615: symlinks/junctions wrong target path on restore