stored: Use logical drive numbers and not drive indexes
Currently the mount protocol between the director and the storage daemon
passes a logical drive number (as seen from the end-user) but the search
algorithm uses the drive index in the autochanger to find the right
drive.

We should number the drives logically and perform the search using that
number. This patch removes the use of any drive index which is a
physical number only needed for external programs that address the drive
via its physical drive index. After this change there is only one place
where the drive index is used and that is in edit_device_codes() which
creates the cmdline for the external programs like mtx-changer etc.

Fixes #634: Storage daemon should use logical drive numbers and not
            drive indexes