Fix: Wrong cookie lifetime
Sets the cookie lifetime to zero to reset session lifetime to it's maximum with every click,
this was set to a wrong value. Also session maximum life- and idle-time is now set to
one hour by default. These settings should be configurable via an ini file in future.
In addition, the session idle-timeout controller plugin is probably no longer needed,
which would require some code cleanup.

Fixes #648: Timeout way too short and other options that need to be configurable