plugins: Enhance plugin interface.
Added new callback method for retrieving the number of
instances of a certain plugin. This mehod was added to all plugins.

This method can be used by the plugin to see if it is the first plugin
being loaded and as such initialize some subsystem that only
should be initialized once per program.

This method can also be used to see if a plugin is the last one being
used and as such needs to call some generic global cleanup hook
for some subsystem.