lib: Use POOL_MEM in ini serialize function.
On master we changed the code from POOLMEM ** to POOLMEM *&
but the new Solaris Studio 12.5beta runs into thye following error.

"ini.c", line 426: Error: Overloading ambiguity between
"ConfigFile::serialize(const char*)" and
"ConfigFile::serialize(char*&)".

We now switched to using POOL_MEM which is a proper class and as such
doesn't give an overloading ambiguity. The POOLMEM type is a typedef of
a char pointer and does with this compiler. Strange thing is that it
does not with any earlier versions of the Solaris studio compiler.