catalog: Lets pass on try_reconnect and exit_on_fatal
When implementing the new try_reconnect and exit_on_fatal flags to
workaround the fact that we handle a database shutdown badly I seem to
have forgotten to pass the arguments in all needed calls. This patch
fixes that.

From now on also most of the flags to instantiate a new connection to
the database need an explicit setting so we catch these kind of problems
in the future much easier. There is only one default left as last
argument to the function to get a new pooled or unpooled connection
which says if it should return a private database connection or not.