rados: Fix restore of objects.
We need to set the correct object_name in createFile() plugin method.
The plugin saves the different objects under a path that looks like
<pool_name>/<object_name> so on restore we only need the last part
of the filename which we store in p_ctx->object_name which is used
in the pluginIO() method to write the data to the correct object.