catalog: Fix argument mixup.
Seems the last fix to the catalog functions was not fully right as we
didn't fix all functions to have the same argument order which leads to
the wrong value being passed on to some functions. This patch fixes that
as from now on need_private is the last argument for each function and
from now on the number of default arguments for db_init_database() is
just like the pooling functions only the need_private flag. We thank
Coverity for finding this nasty bug.