lib: Add support for 16 bits config vars
We need support for the int16_t and uint16_t types in the config parser.
This patch also adds a union in the s_lex_context which holds the values
instead of having a set of separate values. (only union u for single
values and u and u2 for ranges. As only one type is used per scan this
makes the structure a bit smaller in memory)

This also adds support to sscanf for int16_t and uint16_t. e.g. it adds
the %hu and %hd formats to the sscanf implementation. The print functions
already did understand these formats.