dird: Validate pool types in config.
This patch adds a special new config type CFG_TYPE_POOLTYPE which is
used to parse pool types and makes sure only valid pool types are
allowed. Up until now the pool type is stored as a CFG_TYPE_STRNAME and
a check in the database checks if the pool type is really valid. This is
far from user friendly, from now on the DIR will validate the pool type
and store it in the correct way in the config using a filter table which
ignores any end user case sensitivity.

Fixes #664: Validate pool types in config