ndmp: Change NDMP SCSI element status to linked list.
In its current implementation the NDMP SCSI element status uses a fixed
size array with SMC_MAX_ELEMENT(80) elements. This is far from efficient
as we just want to be able to store as many elements as we actually have
without having the additional overhead of unused items. Further more the
element have either a primary_vol_tag or an alternate_vol_tag but in the
old scheme there was a fixed size for both, in the new scheme we only
allocate room for the actual vol_tag that is used.