Fixes LDAP Fileset browsing in restore filetree
This is a workaround which should fix the following bug reported.

bvfs_lsdirs should deliver @LDAP/ if path argument is left empty like it's
giving a / for *nix and all drives for windows clients on top level.

It currently only works if path argument get's an @, like:

*.bvfs_lsdirs jobid=684,691 path=@
282 0 0 0 A A A A A A A A A A A A A A .
282 0 0 0 A A A A A A A A A A A A A A ..
281 0 0 0 A A A A A A A A A A A A A A @LDAP/
282 0 0 0 A A A A A A A A A A A A A A @

Fixes #597: Python LDAP Plugin Fileset not visible in restore