vss: Remove SetBackupSucceeded fallback code.
IVssBackupComponents::SetBackupSucceeded method only needs to be called
when explicit entries are added to the backup snapshotset using the
IVssBackupComponents::AddComponent method. As we only have implicit
entries as we don't call IVssBackupComponents::AddComponent there is no
need to call IVssBackupComponents::SetBackupSucceeded at all.

If we finish our VSS plugin that will handle this part of the VSS API.