vss: Fix IVssBackupComponents::SetBackupState call
We currently always call the IVssBackupComponents::SetBackupState method
with the arguments bSelectComponents and bBackupBootableSystemState set
to true. This indicates that we do a backup or restore operation in
component mode and that a bootable system state backup is being
performed. This is not really true until we fully implement the whole
VSS API in a new dedicated VSS plugin. As such this patch adds a new
event bEventVssSetBackupState which is raised when normally a
IVssBackupComponents::SetBackupState call is done the plugin can
register this event and implement its own call to
IVssBackupComponents::SetBackupState with the options it want to set and
return an bRC_Skip to let the core not call the method.

This patch also adds a new event named bEventVssPrepareForBackup which is
raised just before a call to IVssBackupComponents::PrepareForBackup is
done.

Fixes #562: VSS Plugin is missing method
            IVssBackupComponents::SetBackupSucceeded