Added new default configuration based on directory structure
Added a new default configuration under
  configs/BASE/

It contains a basic configuration for all components as configuration
directories (bareos >= 16.2.4).
It also uses TLS, as we expect that if the test succeeds with TLS it
would proberly also succeed without.

One major advantages of using the configuration subdirectory structure
is that it is now easy to added additional resources to the
configuration of a test.
Tests specific resoures that are defined in
  * configs/$TestName
extends (or overwrites) the configuration defined in configs/BASE.

Tests:
  * tests/tls-client-initiated-connection-backup-test
    * adapted to use the new configuration directory structure

  * tests/config-add-client-test
    * new test, that adds clients via configuration API

  * tests/1-example-test
    * example tests, including comments to give a basis for creating
      additional tests.