dird: Allow to restore from copies.
Until now you can specify the keyword copies when doing a restore
but it will only show the copies but not actually use them for restore.
This is somewhat counter intuitive so this patch adds the code that
asks if you really want to use the copies for the restore. If you
specify the yes keyword you can skip this question.