dird: Simplify temp table.
Currently a lot of fields are inserted into the temp table which are not
used at all. Now that we lookup the storage bits using the new
uar_list_jobs_by_idlist query there is no need to have this overhead
still in the temp table which essentially only needs the JobId,
Termination time and Start time of the Job.