dird: More work to make storage more generic.
We introduce a new type named slot_flags_t for slot
specific flags. This is used in the volume list structure which
allows us to store the more generic slot flags that NDMP will
give back from the SCSI element status e.g. things like if a
slot can be used for import or export or if its excluded from
our view and if if things must be done by operator or can be
done by the robot arm (MTE).

This should make things mostly future proof if we ever move
more of the logic from the SD to the DIR to remove the whole
split brain reservation logic which now exists between the two.