config: Refactor escape_string function.
The escape_string function used for pretty printing from now on has a
POOL_MEM as argument to store the escaped data in. It also now sizes the
buffer right based on the len argument given. This way the calling code
can be cleaned up quite a bit as it now does the same action multiple
times which leads to much more code maintenance.