Restore: Bugfix
An exception has been thrown when no jobids for a specific client were available.

This patch allows an emtpy jobid if no backup jobs are available.

We just drop the jobid argument from isset(), so the argument list is correct
in case of not given jobid, by that the exception is no longer thrown.

Also we no longer try to retrieve related jobids via bvfs api for a not given jobid
which lead to unnecessary load on dird. In this case $jobids, $files and
$directories is immediatly set to null.

Note: bfvs_get_jobids needs at least one jobid to retrieve related jobids for a most
recent restore.