Added MaxFullConsolidations directive
With AlwaysIncrementalMaxFullAge, it is already possible
to minimize the amount of data being copied when
consolidations are done.

MaxFullConsolidations was introduced to be able to
limit the amount of consolidation jobs per run.

This makes it possible to better distribute the load
with many backups to be consolidated, so that e.g. every
day only a maximum of two full consolidations will be run.

The jobs that are skipped using this directive will then
be run the next call of the consolidate job.