bcommand: Fix acl regular expression matching.
We should check for partial matches and disallow those. For that we need
to lose the REG_NOSUB flag and pass a nmatch and pmatch value to
regexec() and check the start and end offset and make sure that its the
full string that matches. We also check if the string has any hints as
to if its a regular expression at all.