bcommand: Add tri-state filter support.
This adds so called tri-state support to the filtering which allows to
return a tri-state (e.g. show, suppress, unknown) which fits better then
a boolean that only can have 2 values. When the filter column is not
filled we cannot really check if a resource is enabled/disabled/exists
or passes an ACL so we now return OF_FILTER_STATE_UNKNOWN indicating
this. We now count all OF_FILTER_STATE_SHOW and OF_FILTER_STATE_UNKNOWN
states returned by the different filters set (set in the calling function)
when all filters are applied and there are no filters returning an
explicit OF_FILTER_STATE_SHOW and we do have one or more
OF_FILTER_STATE_UNKNOWN return values we will explicitly suppress this
value as we cannot be sure that we should or should not show the data.