stored: Status storage should only show configured devices.
When listing the volumes in status storage we still list all volumes on
all devices while we should only list the volumes on the device the
status is asked about.

This also addresses a TODO to move the status function from vol_mgr.c to
status.c as is expressed in the comment in the code but never
implemented as it seems.