cats: Retire sql glueing layer.
When I refactored the catalog backends in 2010 I was talked in to use
some glueing layer so the calling functions didn't need to much
changing. That has served it purpose but now after 6 years its time to
take the next step and move B_DB functions into proper class methods.
This patch removes the old glue layer and extends the B_DB class with
all methods needed to be a proper abstraction of the different database
functions needed. Also all old database functions are now promoted to
class methods. As part of the rewrite we also refactored quite a bit of
code and dropped the db_ prefix where it made sense as its now kind of
obvious its a method doing something with the database when its part of
a B_DB class that handles database abstraction.

I also made all the B_DB members now protected which is a lot easier now
most methods handle the inner content of the class and we don't have to
many external functions poking into our internal structures.

All low level methods previously protected via the bdb_priv.h header
are now proper private methods in the B_DB class and are only visible
to methods in source files that set _BDB_PRIV_INTERFACE_. As such we
control the access to these low level methods to either the methods in
the catalog shared libs or to the actual database backend drivers.

Lets call this catalog refactoring phase 3 after phase 1 and 2 which was
done in 2010.