Required Command ACL validation
This adds a new controller plugin to validate in each controller action
if all required commands to run the module properly are given in the
used Console/Profile Command ACL.

Additionally a new view helper is introduced to centralize the ACL alert
message which is displayed when the given Command ACL does not fit the
requirements.

This commit changes the views and controllers of the following modules:

- Dashboard
- Jobs
- Restore
- Clients
- Schedules
- Storages
- Pools
- Media/Volumes
- Director

Furthermore this commit removes some whitespace from the Director Model.