NDMP: Avoid double slashes in ndmp directory names
The old code added slashes to directories without
checking if the slash is not already there.

This was both done in the code handling NDMP4_FH_ADD_DIR
and in the code handling NDMP4_FH_ADD_FILE calls.

We now we only add slashes if needed, which avoids the
problems caused by double slashes in the filenames.