Prevent from DB lock waits on concurrent use of .bvfs_update
Until this change, the HasCache column in the Job table was either 0 or
1, where 0 represents the state bvfs cache does not exist and 1
represents bvfs cache exists. This change introduces -1 representing the
state bvfs cache update in progress. Before starting the bvfs cache
update, it is now checked if HasCache is -1, in that case the
.bvfs_update jobid=n command will return an ERROR: BVFS reported a
problem for 107. This change also prevents from duplicate key violations
that could occur before on the PathHierarchy table.