Soft Quota: add soft quota check after job run
Before, check_softquota has was called at the beginning of a backup job.
The result was that the Soft Quota Grace Time was not set
when the client soft quota has been exceeded, but when the next job for
this client started.

This does not meet the expectation when using soft quotas.

Now the job report after a backup job also better reflects the real quota
status.