doxygen: introduced doxygen to the project
We introduce the usage of doxygen for the code
documentation.

Therefore existing files were altered so that doxygen
understands the documenting comments:

1. Documentation of functions

Many functions already had a comment block in front of the function
like:
----------------------------------------------------------------
/*
 * This glues the NDMP File Handle DB with internal code.
 */
static inline void register_callback_hooks(struct ndmlog *ixlog)
----------------------------------------------------------------

This was altered so that doxygen understands that this is a documenting
comment, which means that the comments start with '/**' instead of '/*':

----------------------------------------------------------------
/**
 * This glues the NDMP File Handle DB with internal code.
 */
static inline void register_callback_hooks(struct ndmlog *ixlog)
----------------------------------------------------------------

2. Documentation of files

Almost all files already had a description of the file itself like:
----------------------------------------------------------------
/*
 * Backup specific NDMP Data Management Application (DMA) routines
 *
 * Marco van Wieringen, May 2015
 */
----------------------------------------------------------------

This was changed so that the describing part is recognized as file
description by doxygen by adding the @file keyword.
Existing copyright notes were kept in "normal" comment.

----------------------------------------------------------------
/*
 * Marco van Wieringen, May 2015
 */
/**
 * @file
 * Backup specific NDMP Data Management Application (DMA) routines
 */
----------------------------------------------------------------

3. Documentation of struct/class members

Structure and class member description already exists in a lot of
places like:

----------------------------------------------------------------
struct JOB_DBR {
   JobId_t JobId;
   char Job[MAX_NAME_LENGTH];         /* Job unique name */
   char Name[MAX_NAME_LENGTH];        /* Job base name */
----------------------------------------------------------------
As doxygen usually expects the documentation for a certain element
BEFORE the element itself, we need to do two things here:

* we change the /* to /** so that doxygen knows that this is documentation
* we need to add a < to the /** so that we have /**< which tells doxygen
  that this documents the element BEFORE the comment.

----------------------------------------------------------------
struct JOB_DBR {
   JobId_t JobId;
   char Job[MAX_NAME_LENGTH];         /**< Job unique name */
   char Name[MAX_NAME_LENGTH];        /**< Job base name */
----------------------------------------------------------------