B_DB: store the sql queries in std::map
We now have the queries stored in separate files for each
query in src/cats/dml.

Header files are created from those queries and
are compiled in.

During runtime, a std::map is created holding the
query name and query text.

The sql calls use the query_name which is much more readable
than the number that we had before.