Update RestoreModel.php
There's a situation where if $mergefilesets==0 and $mergejobs==1, the $cmd variable isn't set and code is executed, leading into an error trying to get jobids. The user shouldn't be able to select this situation, but I have tried this, and it can happen.
I'm proposing a change in the if-structure so that $cmd is always correctly set.